+++ colors.py
@@ -1,0 +1,6 @@
+# Define the colors we will use in RGB format
+BLACK = (0, 0, 0)
+WHITE = (255, 255, 255)
+BLUE = (0, 0, 255)
+GREEN = (0, 255, 0)
+RED = (255, 0, 0)
